// src/enums/CarEnums.ts
export enum FuelType {
  GASOLINE = "Gasoline",
  DIESEL = "Diesel",
  ELECTRIC = "Electric",
  HYBRID = "Hybrid"
}

export enum CarColor {
  RED = "Red",
  BLUE = "Blue",
  WHITE = "White",
  BLACK = "Black",
  SILVER = "Silver",
  GREEN = "Green"
}

export enum TransmissionType {
  MANUAL = "Manual",
  AUTOMATIC = "Automatic",
  CVT = "CVT"
}